requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

# Providers needed for NASA APOD ETL pipeline
apache-airflow-providers-http>=5.6.0
apache-airflow-providers-postgres>=6.5.0
